Extract `BlobStoreCorruptionUtils`
Makes these utility methods available to other test suites (to be added
in future PRs).

Relates #111954
